requirements.txt
streamlit==1.46.1
gspread==6.2.1
pandas==2.3.0
google-auth==2.23.4
plotly==6.2.0
python-dotenv==1.0.0
